app/backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
pydantic==2.11.9
pydantic-settings==2.4.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.10
alembic==1.16.5
pytest==8.4.2
python-jose[cryptography]>=3.3
bcrypt==4.0.1